feat: split segmentation utils
