+++ src/__init__.py
@@ -2,13 +2,15 @@ from .segment import run_segment
 from .utils import (
     progress_bar,
     remove_file,
+    resample_and_save_nifti,
+    get_resampled_header,
+    align_brain,
+)
+from .segmentation_utils import (
     correct_bias_field,
     fit_intensity_field,
     get_atlas,
-    resample_and_save_nifti,
-    get_resampled_header,
     get_partition,
-    align_brain,
     cleanup,
     get_cerebellum,
 )
